This test was failling when run inside the logstash test.
We were explicitely defining a set amount of calls to `Time.now` (2), but
when we are testing this plugin inside logstash we are doing one more call to `Time`
inside the `Stud.stoppable_sleep` method to adjust the sleep to the skew
of the clock. This PR changes the expectation to `at_least` instead of
`at_most`. The real problem rely in the fact that this plugin uses a
configured pipeline to do the actual testing but this should be
refactored into a simple unit test.

Fixes #52

Fixes #53